requirements.txt
keystone-engine==0.9.2
pyelftools==0.30
